pyproject.toml
[project]
name = "heic2jpg"
version = "0.1.0"
description = "Convert iPhone HEIC photos to high-quality JPG format optimized for silver halide printing"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [
    { name = "HEIC Converter Team" }
]
keywords = ["heic", "jpg", "image", "conversion", "photography", "printing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

dependencies = [
    "pillow>=12.1.1",           # JPG encoding/decoding (CVE-2026-25990 fix)
    "pillow-heif>=0.13.0",      # HEIC decoding
    "piexif>=1.1.3",            # EXIF metadata handling
    "numpy>=1.26.0",            # Image array operations
    "opencv-python>=4.8.0",     # Image analysis
    "rich>=13.0.0",             # CLI progress and formatting
    "click>=8.1.0",             # CLI framework
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.90.0",       # Property-based testing
    "ruff>=0.1.0",              # Linting and formatting
    "mypy>=1.7.0",              # Type checking
    "bandit>=1.7.5",            # Security scanning
]

[project.scripts]
heic2jpg = "heic2jpg.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/heic2jpg"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=heic2jpg",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "property_test: Property-based tests using hypothesis",
    "integration: Integration tests",
    "unit: Unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py314"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",    # Unused function arguments in tests
    "S101",   # Use of assert in tests
]
"src/heic2jpg/models.py" = [
    "TCH003",  # Path is used at runtime in dataclass fields
]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "pillow_heif.*",
    "piexif.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # Allow assert in tests
